Improve compatibility with asdf and rbenv (#11)
Both of these ruby version managers expect a `.ruby-version` file, as
they do not implement a Gemfile parser. By including a `.ruby-version`
file in the project root, it should make it slihtly easier for other
contributors to get started.

That said, it does add another place to change the version number on
language upgrades, which may be an unnecessary or frustrating
constraint.