Rewrite calls to action to be more positive
Some of the critique coming in seems focused on how the tone feels like
one of shame and condemnation instead of one of building shared hope and
solidarity.

This attempts to shift the tone of the next steps to something people
opt into, instead of something they are being shamed into.