Remove resque in favor of sidekiq
Most of the projects I maintain rely on SideKiq and ActiveJob, as they
are both pretty lightweight and fully featured. Since I'll be the one
getting paged when the shit hits the fan, I would prefer to rely on the
libraries I'm comfortable with for background queue processing.

Also, we can use [Heroku
Scheduler](https://devcenter.heroku.com/articles/scheduler) to run jobs
on a particular schedule without requiring a full dyno purely to enqueue
the job.