Remove resque in favor of sidekiq (#21)
Most of the projects I maintain rely on SideKiq and ActiveJob, as they
are both pretty lightweight and fully featured. Since I'll be the one
getting paged when the shit hits the fan, I would prefer to rely on the
libraries I'm comfortable with for background queue processing.

Also, we can use [Heroku
Scheduler](https://devcenter.heroku.com/articles/scheduler) to run jobs
on a particular schedule without requiring a full dyno purely to enqueue
the job.

* Migrate database before releasing to prevent downtime
The [Heroku Release Phase](https://devcenter.heroku.com/articles/release-phase) is a great tool for ensuring the database is up to date or performing any other pre-work when doing a deploy.

* For some reason, Heroku fails if we don't have coffeescript
I would love to purge us of sprockets and webpacker entirely at some point; but I do not have the bandwidth today.